+++ src/LinearRegression.cpp
@@ -51,7 +51,7 @@ Class LinearRegression::classify(Token &token) {
     return clf;
 }
 
-double LinearRegression::classificationProbability(Class c, Token &token) {
+float LinearRegression::classificationProbability(Class c, Token &token) {
     RowVectorXd vec(dataset.dimension() + 1);
     for (int i = 1; i < dataset.dimension()+1; ++i)
         vec[i] = token[i-1];
